Fix texture view synchronization/refcounting;
We were conflating "parent struct of Sync pointer" and "object to be
refcounted", which isn't the case for texture views.

To fix it, add both pointers to the Access struct.  This sucks because
it increases the size from 16 bytes to 24 bytes.

There might be other solutions like a "texture view mask" or having
buffers/textures store a pointer to their sync or something.  But these
have some drawbacks as well.  May revisit in future.