Remove pipeline lock;
This didn't actually make things threadsafe, because you could still
have people reading from the lookup during a draw while it's being
updated in a reset/submit.

Instead, let's just have a rule where you can't do anything during a
submit, and then only update the lookup during submit.

This does mean that pipelines will get "dropped on the floor" if you
issue draws that compile pipelines, but choose to reset the pass instead
of submitting the draws.  In that case, it isn't safe to update the
lookup during the reset, so instead we just wait on the jobs and avoid
updating the map with the pipelines.  I guess this means you could run
into a pathological case where you do this over and over again and run
out of pipelines?  Don't do that.