Refactor buffer allocation;
Pretty much the same, but fixes an obscure lifetime issue with pass
buffer allocation.

If a pass's buffer fills up, it can't be recycled immediately because
that might cause the buffer to get reused too soon if the pass is
submitted multiple times (with draws added in between that exceed 4MB of
temp buffer allocations).

Instead, the strategy is to have the pass keep track of "full" buffers
in a pass-local freelist.

Whenever the pass is submitted, the tick of each full buffer is set to
the current tick, so we know the tick when it's safe to reuse them.

When a pass is reset (or destroyed), all of its full buffers are
returned to the global freelist of stream buffers.  The current one can
remain in use.

Pooling the buffers of all the passes globally like this increases reuse
opportunities but also increases contention.

These changes also serve as preparation to suballocate Buffer objects in
the graphics module, which makes it easier to batch more stuff.