'sample' TextureFeature no longer implies linear filtering support;
Basically it's somewhat common for depth-stencil formats to not support
linear filtering and that is kind of annoying because you can't create
depth textures with the `sample` usage.  Instead we'll just ignore the
LINEAR format feature bit for now.

In the future I'd like to fix this by silently demoting individual
texture's filtering to nearest when linear is not supported for the
format, but this requires per-texture sampler settings which isn't done
yet.