cmake: add ENABLE_EXPORTS on macos/linux/wasm;
Revision 7373bb bumped CMake version from 3.3 to 3.5, and in 3.4 CMake
stopped exporting symbols in executables by default.  lovr relies on
this so ffi can load a few symbols like os_get_glfw_window.

Don't enable it on Android because executables aren't built there, and
don't do it on Windows because it produces an annoying .lib file and
isn't necessary.