Fix lovr.filesystem.setIdentity on Android;
Previously, setting the identity a second time would return false
indicating failure, but the error was ignored by the Lua bindings.
When exceptions were removed in #804, .setIdentity started throwing
errors for this.  This is probably good, except Android sets the
identity at startup, so the .setIdentity call in boot.lua started
throwing.  Let's just change .setIdentity to be an explicit no-op on
Android.