Don't recompute collider mass when it makes Jolt grumpy;
Currently if you add a "static" shape (MeshShape/TerrainShape, shapes
that have undefined mass) to a Collider without one, and the Collider
has automatic mass enabled, Jolt will trigger an assert and say that you
can't compute the mass.

So now instead of always recomputing mass when shapes are changed and
automatic mass is enabled, we'll instead check to see if the collider
has any of those static shapes, and avoid recomputing the mass instead,
to satisfy Jolt.

Note that center of mass will still update.

And as soon as you remove all of the static shapes from the collider,
and it has automatic mass enabled, it should recompute the "real" mass
from the rest of the shapes.