webgpu: don't clear texture atlases;
Currently gpu_write_texture is synchronous instead of async (it doesn't
run on the command buffer timeline).  So clearing the font texture atlas
will wipe out any glyphs that were uploaded in the same frame when the
atlas was created.

WebGPU already clears textures anyway, so we can just avoid clearing the
texture on WebGPU.  This is pretty hacky, should find a better approach.